Split the OU data generator into separate download and merge phases
