+++ .github/workflows/update-standing-data.yml
@@ -31,24 +31,27 @@ jobs:
       - name: Install dependencies
         run: npm install
 
-      - name: Download data files
+      - name: Download offence code data files
         run: npm run download-offence-code-data
 
       - name: Merge offence code data files
         run: npm run merge-offence-data
 
-      - name: Download Organisation Unit data
-        run: npm run download-organisation-unit-data 
+      - name: Download Organisation Unit data files
+        run: npm run download-organisation-unit-data
+
+      - name: Merge Organisation Unit data files
+        run: npm run merge-organisation-unit-data
 
       - name: Report Status
         if: always()
         uses: ravsamhq/notify-slack-action@v1
         with:
           status: ${{ job.status }}
-          notification_title: 'GitHub action failed'
-          message_format: ':elmo-fire: *{workflow}* {status_message} in <{repo_url}|{repo}>'
-          notify_when: 'failure'
-          footer: 'Linked to Repo <{repo_url}|{repo}>'
+          notification_title: "GitHub action failed"
+          message_format: ":elmo-fire: *{workflow}* {status_message} in <{repo_url}|{repo}>"
+          notify_when: "failure"
+          footer: "Linked to Repo <{repo_url}|{repo}>"
         env:
           SLACK_WEBHOOK_URL: ${{ secrets.ACTION_MONITORING_SLACK }}
